fixed problem where sometimes hrv would go as int and others as float

+++ companion-app/app/src/main/java/com/example/dissertationcompanionapp/presentation/viewmodels/MQTTViewModel.kt
@@ -126,7 +126,7 @@ class MQTTViewModel(
             put("measurement","sensors/hrv")
             put("uuid", uuidRepository.getUUID()?:return)
             put("timestamp", timestamp)
-            put("value", hrv)
+            put("value", hrv.toFloat())
         }
 
         mqttClient?.publishWith()
